update types for maxSupply to int64 and initialSupply to uint64 (#59)
* update types for maxSupply to int64 and initialSupply to uint64


* update approveNFT and getApproved to match signature in HederaTokenService and services code


* update decimals to uint32

+++ hts-precompile/HederaTokenService.sol
@@ -154,8 +154,8 @@ abstract contract HederaTokenService is HederaResponseCodes {
     /// @return tokenAddress the created token's address
     function createFungibleToken(
         IHederaTokenService.HederaToken memory token,
-        uint initialTotalSupply,
-        uint decimals) nonEmptyExpiry(token)
+        uint64 initialTotalSupply,
+        uint32 decimals) nonEmptyExpiry(token)
     internal returns (int responseCode, address tokenAddress) {
 
         (bool success, bytes memory result) = precompileAddress.call{value: msg.value}(
@@ -177,8 +177,8 @@ abstract contract HederaTokenService is HederaResponseCodes {
     /// @return tokenAddress the created token's address
     function createFungibleTokenWithCustomFees(
         IHederaTokenService.HederaToken memory token,
-        uint initialTotalSupply,
-        uint decimals,
+        uint64 initialTotalSupply,
+        uint32 decimals,
         IHederaTokenService.FixedFee[] memory fixedFees,
         IHederaTokenService.FractionalFee[] memory fractionalFees) nonEmptyExpiry(token)
     internal returns (int responseCode, address tokenAddress) {
